+++ examples/data-sources/iosxr_evpn_interface/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxr_evpn_interface" "example" {
+  interface_name = "Bundle-Ether12"
+}