+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Add `load_interval` attribute to `iosxr_interface` resource
 - Add `iosxr_evpn_evi` resource and data source
 - Add `iosxr_evpn_group` resource and data source
+- Add `iosxr_evpn_interface` resource and data source
 
 ## 0.1.4